refactor Dashboard into subcomponents
