+++ frontend/src/scenes/dashboard/DashboardHeader.scss
@@ -1,4 +1,4 @@
-.dashboard-selection {
+.dashboard-header {
     display: flex;
     justify-content: space-between;
